refactor(core): extract chart types to dedicated module
Phase 0 of Chart god-class refactoring.

- Extract 15 public types from chart.ts to chartTypes.ts
- Add ChartEventBus in chartContext.ts (lightweight sync event bus)
- Update import paths in controllers/types.ts and createChartController.ts
- Remove unused PaneCapabilities import from chart.ts
- Zero behavioral change; all 956 tests pass
